+++ blobfile/__init__.py
@@ -25,7 +25,6 @@ from blobfile._ops import (
     set_mtime as set_mtime,
     configure as configure,
     BlobFile as BlobFile,
-    scandir_auto_parallel as scandir_auto_parallel,
 )
 from blobfile._context import Context as Context, create_context as create_context
 from blobfile._common import (
@@ -53,7 +52,6 @@ __all__ = [
     "isdir",
     "listdir",
     "scandir",
-    "scandir_auto_parallel",
     "makedirs",
     "remove",
     "rmdir",